Fixes bug 6376: core.thread.thread_scanAll doesn't scan the stack due to ASLR on Mac OS X 10.7
When ASLR is enabled, the stack address of Mac OS X is no longer fixed at 0xc000_0000 -- actually, often higher than this.
Previously, the function rt.memory.rt_stackBottom returns a hard-coded value of 0xc000_0000, which is no longer valid with ASLR.
Functions which rely on this function could no longer work properly.
In particular, core.thread.thread_scanAll cannot scan the main thread's stack since the stack pointer bounds are reversed.
The GC, which marks whether pointers shouldn't be freed yet, will miss all the objects on stack as thread_scanAll is used.
This causes pre-mature deallocation when a GC collection is invoked.
In particular, when throwing an exception (the pointer is on stack), the TraceInfo allocated is > 2 KiB, which will cause a page allocation that invokes the GC.
This makes the exception object be destroyed before checking its type, making it totally uncatchable and ignored, thus previously unreachable statements after the 'throw' will now be reachable, and other undefined behavior will arise.