Fix range violaions in Posix exception handling code.
The code would previously lead to a range violation (and thus an infinite loop because an exception was thrown during exception handling) when druntime was built without bounds checking disabled.

At first sight it seems like the old code could work, but what really happens is that it first index into the (dummy) length 1 static array, and then takes the address, not the other way round.

phi and pcb are intended to be pointers to DHandlerInfo/DCatchBlock, this is also what the original C code did.