Move place of 'import core.sys.posix.sys.mman'.
The 2nd import of module core.sys.posix.sys.mman is moved in an else block. As result, method freeStack uses the same traits as method allocStack.

Background is that the early import creates a reference to the Posix module on Windows. This gives an linker error on Windows if the Posix module is not compiled.