Allow GC implementations to be swappable.
This wraps up GC implementations in a proxy so that they can easily be picked and initialized at runtime before the user program starts according to the config options passed or embedded.

This also adds a secondary "GC implementation" to swap to based on gc_stub. This implementation allows memory to be managed in a manual style using 'new' and 'delete' instead of relying on the GC.